updates.py: rework update check
This replaces how the addon checks for updates. It will still submit
statistics (if enabled), but will not pay attention to the response
with the location and filename of an upgrade file.

Instead, the addon will read releases.json to determine what upgrade
file it should use. Its priority is to first upgrade to the latest
bugfix release within the current update channel (eg 9.2.5 -> 9.2.6
would occur before 9.2.5 -> 10.02). Bugfixes within the same
channel are also the only upgrades that will occur automatically if
enabled.

When upgrading major versions (9.2.x -> 10.0.x), the addon will send
an on screen notification, but will not take any steps to automatically
upgrade.

Note: examples are only examples. The addon will not run on LE 9.2.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>