os_tools: add function to run system commands
This functions runs the given command, waits for it to finish, then
determines what to return to the caller. It is either the command's output,
or None. If the command generates an error, then it also returns None to
signal to the caller it will have no output to work with and take appropriate
action.

With check=True, if the command's return code is not 0, a
subprocess.CalledProcessError exception will be raised. Error information is
sent to the log for troubleshooting. Additional information is available
in debug if the function is suspected to be the source of the error.

FileNotFoundError exception is when the command isn't found to execute.

Reorganize and add some documentation to os_tools while there.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>