config.py: change RPI_CPU_VER to RPI_DEVICE_TYEP and move from oe.py to config.py
RPI_CPU_VER was parsing the revision code to obtain the processor type used by RPi
boards to determine what type of device they were (ex: value for bcm2711 == RPi4 of some type)

This value is used to determine whether to flash the firmware on the device. While
flashing the firmware for an RPi4 or RPi400 is straightforward, flashing on a CM4 is not, so
it has a different process. This makes relying solely on the processor type not ideal.

This changes to relying on the device type as reported in the revision string:
RPi1A/B/B+, Zero, Zero W, RPi2B, RPi3A/B/B+, RPi4, RPi400, etc...

Signed-off-by: Ian Leonard <antonlacon@gmail.com>